+++ core/src/graph/storage/adjacencylist.rs
@@ -334,6 +334,7 @@ impl WriteableGraphStorage for AdjacencyListStorage {
             cyclic: false,
             rooted_tree: true,
             nodes: 0,
+            root_nodes: 0,
             dfs_visit_ratio: 0.0,
         };
 
@@ -370,6 +371,7 @@ impl WriteableGraphStorage for AdjacencyListStorage {
                 }
             }
         }
+        stats.root_nodes = roots.len();
 
         let fan_outs = get_fan_outs(&self.edges);
         let sum_fan_out: usize = fan_outs.iter().sum();